lesson-65-auth/oauth2/globus/requirements.txt
Flask==0.12.2
Flask-SQLAlchemy==2.2
bcrypt
passlib==1.7.1
python-jose==1.3.2
requests-oauthlib==0.8.0

